refactor: split up utils into different modules
