+++ src/cli/data.py
@@ -4,17 +4,16 @@ data_hub.py"""
 import cement
 from deeporigin.data_hub import api
 from deeporigin.exceptions import DeepOriginException
-from deeporigin.utils import (
+from deeporigin.utils.constants import (
     PREFIXES,
     DataType,
-    _ensure_do_folder,
+)
+from deeporigin.utils.core import (
     _print_dict,
     _print_tree,
     _show_json,
     _truncate,
 )
-
-_ensure_do_folder()
 
 
 class DataController(cement.Controller):
