+++ src/variables/types/env_var.py
@@ -5,8 +5,8 @@ import warnings
 
 import dotenv
 import pydantic
+from deeporigin.utils.core import expand_user
 
-from ...utils import expand_user
 from ...warnings import DeepOriginWarning
 from ..base_type import Variable
