+++ src/variables/types/file.py
@@ -3,8 +3,8 @@ import typing
 import warnings
 
 import pydantic
+from deeporigin.utils.core import expand_user
 
-from ...utils import expand_user
 from ...warnings import DeepOriginWarning
 from ..base_type import Variable
