+++ src/variables/types/git_http_credentials.py
@@ -5,8 +5,8 @@ import warnings
 
 import pydantic
 import validators
+from deeporigin.utils.core import expand_user
 
-from ...utils import expand_user
 from ...warnings import DeepOriginWarning
 from ..base_type import Variable
